Update darwin kernel misc

+++ source/core/arch/CpuFeature.h
@@ -1,11 +1,11 @@
 #ifndef CORE_ARCH_CPU_FEATURE_H
 #define CORE_ARCH_CPU_FEATURE_H
 
-#include "common/headers/common_header.h"
+#include "common_header.h"
 
 class CpuFeatures {
 private:
-  static void FlushICache(void *start, int size) {
+  static void FlushICache(void *start, size_t size) {
     ClearCache(start, (void *)((addr_t)start + size));
   }
